+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Change Log
 
 ### Additions/Changes
 
+- Migrate document drawer to sidebar style.
 - Improve to display gear icon in menu bar while executing a script.
 - Add hidden “Reveal in Finder” menu item to syntax style action menu in Preferences (visible with `Option` key).
 - Improve find result panel behavior with Spaces.